% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{const_fun}
\alias{const_fun}
\title{Calculate sum of internal references}
\usage{
const_fun(
  omega_p,
  rho_std,
  db_inv_std,
  alpha_std,
  beta_std,
  rho_tgt,
  db_inv_tgt,
  alpha_tgt,
  beta_tgt
)
}
\description{
Calculate sum of internal references
}
