% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{const_inv_std}
\alias{const_inv_std}
\title{Dependence of rho parameters on constant in case of only p=1
(H function, Heller, 2021)}
\usage{
const_inv_std(rho_sbl, param)
}
\description{
Dependence of rho parameters on constant in case of only p=1
(H function, Heller, 2021)
}
