% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{const_inv_tgt}
\alias{const_inv_tgt}
\title{Calcuate rho_ltob given const_lb and rho_bfroml
Dependence of rho parameters on constant in case of only p=1
(H function, Heller, 2021)}
\usage{
const_inv_tgt(rho_clb, param)
}
\arguments{
\item{rho_clb}{A intensity in luminance unit (dB Lambert) representing the
internal brightness reference for the target}

\item{param}{A vector of parameter values for the global psychophysical
model containing const_lb, omega_1, alpha_l, beta_l, and beta_b}
}
\value{
A number representing the internal loudness reference for the
standard
}
\description{
Calcuate rho_ltob given const_lb and rho_bfroml
Dependence of rho parameters on constant in case of only p=1
(H function, Heller, 2021)
}
