% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_disp}
\alias{db_disp}
\title{Convert displacement in mm to dB}
\usage{
db_disp(x_t)
}
\arguments{
\item{x_t}{A number representing a displacement of the tactor in mm.}
}
\value{
A number representing the respective sound in dB relative
to 0.001578822 mm displacement.
}
\description{
Convert displacement in mm to dB
}
