% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_inv_disp}
\alias{db_inv_disp}
\title{Convert dB to displacement in mm}
\usage{
db_inv_disp(db_t)
}
\arguments{
\item{db_t}{A number representing a puls in dB.}
}
\value{
A number representing the displacement in mm.
}
\description{
Convert dB to displacement in mm
}
