% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_inv_lambert}
\alias{db_inv_lambert}
\title{Convert dB lambert to cd/m2}
\usage{
db_inv_lambert(db_b)
}
\arguments{
\item{db_b}{A number representing a luminance in dB Lambert.}
}
\value{
A number representing a luminance in cd/m2.
}
\description{
Convert dB lambert to cd/m2
}
