% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_inv_spl}
\alias{db_inv_spl}
\title{Convert dB SPL to Pascal}
\usage{
db_inv_spl(db_l)
}
\arguments{
\item{db_l}{A number representing a sound in dB SPL.}
}
\value{
A number representing the sound pressure in Pascal.
}
\description{
Convert dB SPL to Pascal
}
