% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_lambert}
\alias{db_lambert}
\title{Convert cd/m2 to dB lambert}
\usage{
db_lambert(x_b)
}
\arguments{
\item{x_b}{A number representing a luminance in cd/m2.}
}
\value{
A number representing the respective luminance in dB Lambert
relative to 10^-10 Lambert.
}
\description{
Convert cd/m2 to dB lambert
}
