% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{db_spl}
\alias{db_spl}
\title{Convert sound pressure in Pascal to dB SPL re 20 muPa}
\usage{
db_spl(x_l)
}
\arguments{
\item{x_l}{A number representing a sound pressure in Pascal.}
}
\value{
A number representing the respective sound in dB SPL relative
to 20 muPa.
}
\description{
Convert sound pressure in Pascal to dB SPL re 20 muPa
}
