% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate parameters of global psychophysics model}
\usage{
estimate(data, ntrials, references = c("dependent", "independent", "constant"))
}
\arguments{
\item{data}{A data frame containing the standard and target intensity in dB,
and task direction (bright_loud or loud_bright). Must be sorted loud_bright
first and then bright_loud.}

\item{ntrials}{number of trials per standard.}
}
\description{
Estimate parameters of global psychophysics model
}
\examples{
data(matching)
datlist <- make_datlist(data = matching[matching$id == 2, ], 48)
m <- estimate(data = matching[matching$id == 2, ], 48, references = "constant")
print(m, pars = c("alpha_l", "beta_l", "beta_b", "omega1", "omega", "const_lb",
                  "const_bl"), probs = c(.025, .975))

}
