% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gpm_multiple_p}
\alias{gpm_multiple_p}
\title{Calculate global psychophysics model prediction}
\usage{
gpm_multiple_p(
  standard_intensity,
  alpha_std,
  alpha_tgt,
  beta_std,
  beta_tgt,
  p,
  omega_1 = NULL,
  omega = NULL,
  omega_p = NULL,
  rho_std = NULL,
  rho_tgt = NULL,
  const = NULL,
  task = c("bright_loud", "loud_bright")
)
}
\arguments{
\item{standard_intensity}{A number or vector representing the physical
intensity/ies of the standard stimulus/stimuli (in dB).}

\item{alpha_std}{A number for the parameter of the psychophysical function
of the standard.}

\item{alpha_tgt}{A number for the parameter of the psychophysical function
of the target.}

\item{beta_std}{A number for the parameter of the psychophysical function
of the standard.}

\item{beta_tgt}{A number for the parameter of the psychophysical function
of the target.}

\item{p}{An integer (1, 2, or 3) representing the production ratio to be used.}

\item{omega_1}{A number representing the weight parameter for the weighting
function.}

\item{omega}{A number representing the exponent in the weighting function.
Default is 0.6.}

\item{omega_p}{A direct specification of the weighted production ratio.
If omega_1 and omega are given, this will be ignored.}

\item{rho_std}{A number for the internal reference parameter pertaining to
the standard dimension.}

\item{rho_tgt}{A number for the internal reference parameter pertaining to
the target dimension.}

\item{const}{A number representing the sum of rho's in the restricted model
for matching / p=1 only.}

\item{task}{Either "bright_loud" or "loud_bright" indicating the task.}
}
\value{
A number or vector representing the physical intensity predicted by
the global psychophysics model, with names reflecting the production
ratio used.
}
\description{
This function extends the global psychophysics model function (gpm) by
allowing calculation of model predictions for a specific production ratio
(p). It handles the weighting calculation internally and provides appropriate
naming of results based on the production ratio used.
}
\examples{
# Calculate prediction for brightness-to-loudness with production ratio p=2
gpm_multiple_p(standard_intensity = 70,
               alpha_std = 1, alpha_tgt = 0.5,
               beta_std = 0.33, beta_tgt = 0.67,
               rho_std = 20, rho_tgt = 20,
               p = 2, omega_1 = 1, omega = 0.6,
               task = "bright_loud")
}
