% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{make_datlist}
\alias{make_datlist}
\title{Create a list of data information to be given to rstan::stan()}
\usage{
make_datlist(data, ntrials)
}
\arguments{
\item{data}{A data frame containing columns standard and match in dB, and
task direction (bright_loud or loud_bright).}

\item{ntrials}{number of trials per standard.}
}
\description{
Create a list of data information to be given to rstan::stan()
}
