% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{predict_gpm}
\alias{predict_gpm}
\title{Predict magnitude productions from global psychophysics model}
\usage{
predict_gpm(standards, task, param, p = 1)
}
\arguments{
\item{standards}{A vector of numbers.}

\item{task}{A character string, either "bright_loud" or "loud_bright"
indicating the task.}

\item{param}{A named vector of numbers representing the model parameters.}
}
\value{
A vector of predicted values.
}
\description{
Predict magnitude productions from global psychophysics model
}
\examples{
param <- create_param_set()
y_bright <- predict_gpm(c(34, 43, 52, 70), "loud_bright", param)
y_loud <- predict_gpm(c(69, 73, 77, 85), "bright_loud", param)
}
