% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{psi_inv}
\alias{psi_inv}
\title{Inverse of psychophysical function phi(x) = alpha * x^beta}
\usage{
psi_inv(x, alpha, beta)
}
\arguments{
\item{x}{A number representing a perceived intentsity (e.g.
brightness or loudness)}

\item{alpha}{A number.}

\item{beta}{A number.}
}
\value{
A number representing a physical stimulus intentsity (e.g.
luminance in cd/m2 or sound pressure in Pascal).
}
\description{
Inverse of psychophysical function phi(x) = alpha * x^beta
}
