% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gpm}
\alias{gpm}
\title{Calculate global psychophysics model prediction of the physical intensity
produced in a magnitude production task with one production ratio}
\usage{
gpm(
  standard_intensity,
  alpha_std,
  alpha_tgt,
  beta_std,
  beta_tgt,
  omega_p,
  rho_std = NULL,
  rho_tgt = NULL,
  const = NULL,
  task = c("bright_loud", "loud_bright")
)
}
\arguments{
\item{standard_intensity}{A number or vector representing the physical
intensity/ies of the standard stimulus / stimuli (in dB).}

\item{alpha_std}{A number for the parameter of the psychophysical function
of the standard.}

\item{alpha_tgt}{A number for the parameter of the psychophysical function
of the target.}

\item{beta_std}{A number for the parameter of the psychophysical function
of the standard.}

\item{beta_tgt}{A number for the parameter of the psychophysical function
of the target.}

\item{omega_p}{A number representing the weighted production ratios W(p).}

\item{rho_std}{A number for the internal reference parameter pertaining to
the standard dimension.}

\item{rho_tgt}{A number for the internal reference parameter pertaining to
the target dimension.}

\item{const}{A number representing the sum of rho's in the restricted model
for matching / p=1 only.}

\item{task}{Either "bright_loud" or "loud_bright" indicating the task.}
}
\value{
A number or vector representing the physical intensity predicted by
the global psychophysics model.
}
\description{
Calculate global psychophysics model prediction of the physical intensity
produced in a magnitude production task with one production ratio
}
